# docs/tech_specs/api-spec.yaml
openapi: 3.0.3
info:
  title: Workflow Orchestration API
  version: 1.0.0
  description: >
    Backend APIs for planning, creating, testing, activating, and observing AI-planned
    workflows that execute via Composio (MCP/SDK). Workflows are represented as DSL JSON.
servers:
  - url: https://api.yourdomain.com
...
paths:
...
  /runs:
    get:
      tags: [Runs]
      summary: List runs (paged)
      description: Returns recent runs, optionally filtered by workflow_id.
      operationId: listRuns
      parameters:
        - in: query
          name: workflow_id
          schema:
            type: string
          description: Filter by workflow id
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Runs page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunsListResponse'
  /runs/{run_id}:
    get:
      tags: [Runs]
      summary: Get run details
      description: Returns run status, step history, and outputs/errors per step.
      operationId: getRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetailsResponse'
        '404':
          description: Not found
  /runs/stream:
    get:
      tags: [Runs]
      summary: Server-Sent Events stream for live run updates
      description: >
        Emits `text/event-stream` updates for runs of a workflow (or a specific run if provided).
        Events include step start/finish, partial outputs (for streaming steps), errors, and completion.
      operationId: streamRuns
      parameters:
        - in: query
          name: workflow_id
          schema:
            type: string
          description: Optional workflow id to subscribe to
        - in: query
          name: run_id
          schema:
            type: string
          description: Optional specific run id
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
                example: "event: step\ndata: {\"run_id\":\"run_123\",\"step_id\":\"slack_post\",\"status\":\"ok\"}\n\n"
  /catalog:
    get:
      tags: [Catalog]
      summary: Get tool catalog snapshot
      description: Returns cached provider/action/param metadata for requested providers (or all).
      operationId: getCatalog
      parameters:
        - in: query
          name: providers
          description: Comma-separated provider ids (e.g., composio/slack,composio/gmail)
          schema:
            type: string
      responses:
        '200':
          description: Catalog snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
...
  /catalog/tools:
    get:
      tags: [Catalog]
      summary: Search tools
      description: Search for tools across all providers or within a specific provider.
      operationId: searchTools
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search query for tool names
        - in: query
          name: provider_slug
          schema:
            type: string
          description: Filter by specific provider slug
        - in: query
          name: tool_type
          schema:
            type: string
            enum: [action, trigger, both]
          description: Filter by tool type
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
          description: Maximum number of results
        - in: query
          name: force_refresh
          schema:
            type: boolean
            default: false
          description: Force refresh from external sources
      responses:
        '200':
          description: Tool search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolSearchResponse'
  /catalog/tools/{tool_name}:
    get:
      tags: [Catalog]
      summary: Get tool by name
      description: Get a specific tool by name, optionally filtered by provider.
      operationId: getToolByName
      parameters:
        - in: path
          name: tool_name
          required: true
          schema:
            type: string
          description: Tool name (e.g., 'chat.postMessage')
        - in: query
          name: provider_slug
          schema:
            type: string
          description: Filter by specific provider slug
        - in: query
          name: force_refresh
          schema:
            type: boolean
            default: false
          description: Force refresh from external sources
      responses:
        '200':
          description: Tool information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        '404':
          description: Tool not found
components:
  parameters:
...
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        type: string
      description: Run identifier
  schemas:
...
    RunsListResponse:
      type: object
      required: [runs]
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/RunSummary'
        next_page:
          type: integer
          nullable: true
    RunSummary:
      type: object
      required: [run_id, workflow_id, status, started_at]
      properties:
        run_id:
          type: string
        workflow_id:
          type: string
        status:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
    RunDetailsResponse:
      type: object
      required: [run_id, workflow_id, status, steps]
      properties:
        run_id:
          type: string
        workflow_id:
          type: string
        status:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        steps:
          type: array
          items:
            $ref: '#/components/schemas/RunStep'
        outputs:
          type: object
          additionalProperties: true
          description: Optional aggregate outputs
    RunStep:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [queued, running, ok, failed, skipped]
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        output:
          type: object
          additionalProperties: true
          nullable: true
        error:
          $ref: '#/components/schemas/StepError'
    StepError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    # ----- Catalog -----
    CatalogResponse:
      type: object
      properties:
        providers:
          type: array
          items:
            $ref: '#/components/schemas/Provider'
    Provider:
      type: object
      required: [id]
      properties:
        id:
          type: string
          example: composio/slack
        version:
          type: string
        actions:
          type: array
          items:
            $ref: '#/components/schemas/ProviderAction'
        triggers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderTrigger'
        rate_limits:
          type: object
          additionalProperties: true
    ProviderAction:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: chat.postMessage
        params:
          type: array
          items:
            $ref: '#/components/schemas/ParamSpec'
    ProviderTrigger:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: message.received
        delivery:
          type: array
          items:
            type: string
            enum: [webhook, poll, schedule]
        params:
          type: array
          items:
            $ref: '#/components/schemas/ParamSpec'
    ParamSpec:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        required:
          type: boolean
          default: false
        enum:
          type: array
          items:
            type: string
    ToolResponse:
      type: object
      properties:
        tool:
          $ref: '#/components/schemas/Tool'
        provider:
          $ref: '#/components/schemas/ProviderInfo'
    ToolSearchResponse:
      type: object
      properties:
        query:
          type: string
          description: Search query used
        provider_filter:
          type: string
          description: Provider slug filter applied
        tool_type_filter:
          type: string
          description: Tool type filter applied
        total_tools:
          type: integer
          description: Total number of tools found
        total_providers:
          type: integer
          description: Total number of providers with matching tools
        results:
          type: array
          items:
            $ref: '#/components/schemas/ProviderTools'
    Tool:
      type: object
      required: [name, display_name, tool_type, version]
      properties:
        name:
          type: string
          example: chat.postMessage
        display_name:
          type: string
          example: Send Message
        description:
          type: string
        tool_type:
          type: string
          enum: [action, trigger, both]
        version:
          type: string
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/ToolParameter'
    ToolParameter:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        display_name:
          type: string
        description:
          type: string
        type:
          type: string
        required:
          type: boolean
          default: false
        default:
          type: string
        validation:
          type: object
          additionalProperties: true
    ProviderInfo:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
          example: slack
        name:
          type: string
          example: Slack
    ProviderTools:
      type: object
      required: [provider, tools]
      properties:
        provider:
          $ref: '#/components/schemas/ProviderInfo'
        tools:
          type: array
          items:
            $ref: '#/components/schemas/Tool'

    # ----- DSLs & Compile/Lint -----
...